updating scripts

+++ dumper/performancePlots/makePlots.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
 
-versions=$1
 
-for version in $versions;do
+for version in V23 V22_b V21_b V20_b;do
     cd lists
     ./makeLists.sh $version
     cd -
@@ -14,32 +13,47 @@ for version in $versions;do
 for dataset in  HggRelVal DYToLLRelVal ZEERelVal; do
     if [ "$dataset" == "DYToLLRelVal" ]
     then
-	./tmp/runCreateHistos2 lists/$dataset"_"$version.list $dataset"_"$version.root   
-	./tmp/runCreateHistos2 lists/$dataset"_noPU_"$version.list $dataset"_noPU_"$version.root
+#	./tmp/runCreateHistos2 lists/$dataset"_"$version.list $dataset"_"$version.root   
+#	./tmp/runCreateHistos2 lists/$dataset"_noPU_"$version.list $dataset"_noPU_"$version.root
     echo "DY ok"
     fi
 
     if [ "$dataset" == "ZEERelVal" ]
     then
-	./tmp/runCreateHistos2 lists/$dataset"_"$version.list $dataset"_"$version.root   
-	./tmp/runCreateHistos2 lists/$dataset"_noPU_"$version.list $dataset"_noPU_"$version.root
-    echo "DY ok"
+#	./tmp/runCreateHistos2 lists/$dataset"_"$version.list $dataset"_"$version.root   
+#	./tmp/runCreateHistos2 lists/$dataset"_noPU_"$version.list $dataset"_noPU_"$version.root
+    echo "ZEE ok"
     fi
 
 
     if [ "$dataset" == "HggRelVal" ]    
     then
-	./tmp/runCreateHistosPhotons lists/$dataset"_"$version.list $dataset"_"$version.root   
-	./tmp/runCreateHistosPhotons lists/$dataset"_noPU_"$version.list $dataset"_noPU_"$version.root
+#	./tmp/runCreateHistosPhotons lists/$dataset"_"$version.list $dataset"_"$version.root   
+#	./tmp/runCreateHistosPhotons lists/$dataset"_noPU_"$version.list $dataset"_noPU_"$version.root
     echo "Hgg ok" 
     fi
 
-    rm plots/h*
-   ./tmp/plotHistos $dataset"_noPU_"$version.root $dataset"_"$version.root plots_$dataset"_"$version.root
-    mkdir ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset
-    cp plots/h* ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/
-    cp ~/www/plots/shashlikUpgrade/index.php  ~/www/plots/shashlikUpgrade/$dateDir/$version/index.php
-    cp ~/www/plots/shashlikUpgrade/index.php  ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/index.php
+    n=`cat lists/$dataset"_"$version.list |wc -l`
+    echo "plotting"
+    if [ $n -ne 0 ]
+    then
+	rm plots/h*
+	./tmp/plotHistos $dataset"_noPU_"$version.root $dataset"_"$version.root plots_$dataset"_"$version.root
+	mkdir ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset
+	cp plots/h* ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/
+	cp ~/www/plots/shashlikUpgrade/index.php  ~/www/plots/shashlikUpgrade/$dateDir/$version/index.php
+	cp ~/www/plots/shashlikUpgrade/index.php  ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/index.php
+
+	rm plots_splitted/h*
+	./tmp/plotDistributions $dataset"_noPU_"$version.root plots_splitted_$dataset"_"$version
+	mkdir ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/
+	mkdir ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/noPU/
+	cp plots_splitted/h* ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/noPU/
+	cp ~/www/plots/shashlikUpgrade/index.php  ~/www/plots/shashlikUpgrade/$dateDir/$version/index.php
+	cp ~/www/plots/shashlikUpgrade/index.php  ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/index.php
+	cp ~/www/plots/shashlikUpgrade/index.php  ~/www/plots/shashlikUpgrade/$dateDir/$version/$dataset/noPU/index.php
+
+    fi
 done
 
 done
